entrega_4/src/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
asttokens==2.4.0
attrs==21.4.0
autoray==0.6.8
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
Choco==1.0.5
circ==0.0.1
cirq==1.3.0
cirq-aqt==1.3.0
cirq-core==1.3.0
cirq-ft==1.3.0
cirq-google==1.3.0
cirq-ionq==1.3.0
cirq-pasqal==1.3.0
cirq-rigetti==1.3.0
cirq-web==1.3.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
comm==0.1.4
contourpy==1.1.1
cotengra==0.5.6
cryptography==42.0.2
cycler==0.12.1
Cython==3.0.8
cytoolz==0.12.3
data==0.4
debugpy==1.8.0
deco==0.6.3
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.7
distlib==0.3.8
distro==1.8.0
docopt==0.6.2
duet==0.2.9
et-xmlfile==1.1.0
executing==2.0.0
fastjsonschema==2.19.1
fibheap==0.2.1
filelock==3.13.1
fiona==1.9.5
fire==0.5.0
Flask==3.0.1
fonttools==4.43.1
frozenlist==1.4.1
funcsigs==1.0.2
future==1.0.0
geopandas==0.14.3
google-api-core==2.17.1
google-auth==2.28.1
googleapis-common-protos==1.62.0
grpcio==1.62.0
grpcio-status==1.62.0
gunicorn==21.2.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
ibm-cloud-sdk-core==3.19.2
ibm-platform-services==0.51.2
idna==3.6
iniconfig==2.0.0
ipykernel==6.27.1
ipython==8.12.3
ipywidgets==8.1.1
iso8601==1.1.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.17.3
jupyter_client==8.6.0
jupyter_core==5.5.1
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
lark==0.11.3
latex==0.7.0
llvmlite==0.42.0
MarkupSafe==2.1.4
matplotlib==3.8.0
matplotlib-inline==0.1.6
maven==0.0.6
mistune==3.0.2
moody==0.2.3
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.4
nbclient==0.9.0
nbconvert==7.16.1
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
nose==1.3.7
numba==0.59.0
numpy==1.26.0
openai==1.6.0
openpyxl==3.1.2
opt-einsum==3.3.0
ordered-set==4.1.0
packaging==23.2
pandas==2.1.1
pandocfilters==1.5.1
parso==0.8.3
pbr==5.11.1
pdfminer.six==20221105
pdfplumber==0.10.3
pickleshare==0.7.5
Pillow==10.0.1
pipdeptree==2.15.1
pipreqs==0.5.0
platformdirs==4.1.0
pluggy==1.4.0
ply==3.11
prompt-toolkit==3.0.39
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.14
pydantic_core==2.14.5
pydot==1.4.2
Pygments==2.16.1
PyJWT==2.8.0
PyLaTeX==1.3.4
pylatexenc==2.10
PyMuPDF==1.23.20
PyMuPDFb==1.23.9
pyparsing==3.1.1
pypdfium2==4.26.0
pyproj==3.6.1
pyquil==3.5.4
pyrsistent==0.20.0
pyspnego==0.10.2
pytest==8.0.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-rapidjson==1.16
pytz==2023.3.post1
pywin32==306
pyzmq==25.1.2
qcs-api-client==0.21.6
qiskit==1.0.1
qiskit-aer==0.13.3
qiskit-ibm-runtime==0.20.0
qiskit-terra==0.25.2.1
quimb==1.7.3
qutip==4.7.5
requests==2.31.0
requests-ntlm==1.2.0
retrying==1.3.4
rfc3339==6.2
rfc3986==1.5.0
rpcq==3.11.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
rustworkx==0.14.1
scipy==1.11.3
seaborn==0.13.0
sh==2.0.6
shapely==2.0.3
shutilwhich==1.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sspilib==0.1.0
stack-data==0.6.3
stevedore==5.1.0
svgwrite==1.4.3
symengine==0.11.0
sympy==1.12
tabulate==0.9.0
tempdir==0.7.1
tenacity==8.2.3
termcolor==2.4.0
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.1
traitlets==5.11.2
types-Deprecated==1.2.9.20240106
types-python-dateutil==2.8.19.20240106
types-retry==0.9.9.4
typing_extensions==4.9.0
tzdata==2023.3
Unidecode==1.3.7
urllib3==2.1.0
virtualenv==20.25.0
wcwidth==0.2.8
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wrapt==1.16.0
xlrd==2.0.1
yarg==0.1.9
yarl==1.9.4
youtube-dl==2021.12.17
